requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.8
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-storages==1.14.6
django-timezone-field==7.1
django-allauth==65.13.1
django-celery==3.1.17
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
exceptiongroup==1.3.1
inflection==0.5.1
kombu==5.6.1
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
